lab-03/requirements.txt
markdown-it-py
mdurl
numpy
pillow
Pygments
rich
shapely
stopit
types-Pillow
setuptools